"Update LoginForm to include Link to Sign Up page and fix typo in RegisterForm"
This commit updates the `LoginForm` component to include a link to the Sign Up page, and fixes a typo in the `RegisterForm` component. The `LoginForm` component now includes a new line with a link to the Sign Up page, allowing users to easily navigate to the Sign Up page if they don't have an account. The `RegisterForm` component had a typo in the text of the link to the Sign In page, which has been corrected.