Update dependencies, add sweetalert2 and react-spinners, update Loading component, and modify LoginForm and EventDetailsPage to use sweetalert2 for error and success messages
The diff includes updates to package-lock.json and package.json for installing new dependencies sweetalert2 and react-spinners. The Loading component is updated to use a ClockLoader from react-spinners. The LoginForm component is modified to use sweetalert2 for error messages instead of console.error and to show a success message using sweetalert2. The EventDetailsPage component is updated to use sweetalert2 for success and error messages.